move range to its own files.  fix test panic.
